Training model generation fixes + v6 : Win Percentages: V6: 68.5%  |  Draw: 0.9%  |  V5: 30.6%

+++ models/game/types.go
@@ -9,8 +9,8 @@ const (
 )
 
 type Position struct {
-	Row int
-	Col int
+	Row int8
+	Col int8
 }
 
 type Board [8][8]Piece
